refactor the action creators to better slipt them up
